requirements.txt
langchain
langchain-community 
langchain-core 
langchain-ollama 
chromadb 
sentence-transformers 
pypdf 
python-dotenv 
unstructured[pdf]
tiktoken